complete?????

+++ 42makefile.sh
@@ -54,12 +54,17 @@ function ft_folder_querry()
 		TLEN=1
 	fi
 	mkdir $1
-	echo -en "$2_FILE" >> $FILE
-	printf "%1s" | tr " " "\t" >> $FILE
-	echo -e "=\t\n" >> $FILE
 	echo -en "$2_DIR" >> $FILE
 	printf "%${TLEN}s" | tr " " "\t" >> $FILE
 	echo -e "=\t./$1" >> $FILE
+	echo -en "$2_FILE" >> $FILE
+	printf "%1s" | tr " " "\t" >> $FILE
+	if [ "$1" == "obj" ]; then
+		echo -en "=\t" >> $FILE
+		echo "\$(patsubst %.c, %.o, \$(SRC_FILE))" >> $FILE
+	else
+		echo -e "=\t\n" >> $FILE
+	fi
 	echo -en "$3" >> $FILE
 	printf "%${TLEN}s" | tr " " "\t" >> $FILE
 	echo -en "=\t" >> $FILE
@@ -87,7 +92,7 @@ function ft_type_querry()
 
 function ft_ar_querry()
 {
-	echo -en "\033[32;1m?\033[0m \033[2m is this project for making a library ? (y/n)\033[0m : "
+	echo -en "\033[32;1m?\033[0m \033[2m is this project for making a library (.a file) ? (y/n)\033[0m : "
 	read LIB
 	if [ "$LIB" == "y" ]; then
 		LIB=1
@@ -105,8 +110,6 @@ function ft_phony_querry()
 	fi
 }
 
-
-
 ft_ar_querry
 ft_member_querry NAME "42" LIB
 echo -en "\n" >> $FILE
@@ -115,23 +118,67 @@ ft_member_querry CCFLAG "-Wall -Wextra -Werror" 0
 echo -en "\n" >> $FILE
 ft_type_querry
 
-
 if [ $TYPE == 0 ]; then
 	echo -en "all\t\t:\t\$(NAME)\n\n" >> $FILE
-	echo -en "\$(NAME)\t:\t\$(OBJS)\n" >> $FILE
 	if [ $LIB == 1 ]; then
-		echo -en "\t\tar rcs \$(NAME) \$(OBJECTS)\n\n" >> $FILE
+		echo -en "\$(NAME)\t:\t\$(OBJECTS)\n" >> $FILE
+		echo -en "\t\tar rcs \$(NAME) \$(OBJECTS)\n" >> $FILE
+		echo -en "\t\t" >> $FILE
+		echo "echo \"\033[0;92mlibrary (.a) file was created\"" >> $FILE
+		echo -en "\n\n" >> $FILE
 	else
-		echo -en "\t\t\$(CC) \$(CCFLAG) \$(HEADERS)\n\n" >> $FILE
+		echo -en "\$(NAME)\t:\t\$(OBJECTS)\n" >> $FILE
+		echo -en "\t\t\$(CC) \$(CCFLAG) \$(HEADERS) \$(OBJECTS) -o \$(NAME)\n" >> $FILE
+		echo -en "\t\t" >> $FILE
+		echo "echo \"\033[0;92mprogram file was created\"" >> $FILE
+		echo -en "\n\n" >> $FILE
 	fi
-	echo -en "clean\t:\n\t\trm -rf \$(OBJECTS)\n\n" >> $FILE
-	echo -en "fclean\t:\tclean\n\t\trm -rf \$(NAME)\n\n" >> $FILE
-	echo -en "re\t\t:\tfclean \$(NAME)\n\n\n" >> $FILE
+	echo -en "clean\t:\n\t\trm -rf \$(OBJECTS)\n" >> $FILE
+	echo -en "\t\t" >> $FILE
+	echo "echo \"\033[0;91mobject files was deleted\"" >> $FILE
+	echo -en "\n\nfclean\t:\tclean\n\t\trm -rf \$(NAME)\n" >> $FILE
+	echo -en "\t\t" >> $FILE
+	echo "echo \"\033[0;91m\$(NAME) was deleted\"" >> $FILE
+	echo -en "\n\nre\t\t:\tfclean \$(NAME)\n\n\n" >> $FILE
+else
+	echo -en "all\t\t:\t\$(NAME)\n\n" >> $FILE
+	echo -en "\$(NAME)\t:\t\$(OBJECTS)\n" >> $FILE
+	if [ $LIB == 1 ]; then
+		echo -en "\$(NAME)\t:\t\$(OBJECTS)\n" >> $FILE
+		echo -en "\t\tar rcs \$(NAME) \$(OBJECTS)\n\n" >> $FILE
+		echo -en "\t\t" >> $FILE
+		echo "echo \"\033[0;92mlibrary (.a) file was created\"" >> $FILE
+		echo -en "\n\n" >> $FILE
+	else
+		echo -en "\$(NAME)\t:\t\$(OBJECTS) \$(HEADERS)\n" >> $FILE
+		echo -en "\t\t\$(CC) \$(CCFLAG) \$(HEADERS) \$(OBJECTS) -o \$(NAME)\n" >> $FILE
+		echo -en "\t\t" >> $FILE
+		echo "echo \"\033[0;92mprogram file was created\"" >> $FILE
+		echo -en "\n\n" >> $FILE
+	fi
+	echo -en "\$(OBJ_DIR)%.o\t:\t\$(SRC_DIR)%.c\t\$(HEADERS)\n" >> $FILE
+	echo -en "\t\t@\$(CC) \$(CCFLAG) -c -I\$(HDR_DIR) \$< -o \$@" >> $FILE
+	echo -en "clean\t:\n\t\trm -rf \$(OBJ_DIR)\n" >> $FILE
+	echo -en "\t\t" >> $FILE
+	echo -n "echo \"\033[0;91mobject files was deleted\"" >> $FILE
+	echo -en "\n\nfclean\t:\tclean\n\t\trm -rf \$(NAME)\n" >> $FILE
+	echo -en "\t\t" >> $FILE
+	echo "echo \"\033[0;91m\$(NAME) was deleted\"" >> $FILE
+	echo -en "\n\nre\t\t:\tfclean \$(NAME)\n\n\n" >> $FILE
 fi
-
-
-
 
 ft_phony_querry
 
-mv .mf Makefile
+if [ -f "./Makefile" ]; then
+	echo -en "\033[32;1m?\033[0m \033[2;91m Makefile is already existent, would you like to overwrite it? (y/n)\033[0m : "
+	read CHECK
+	if [ "${CHECK}" == "y" ]; then
+		chmod 755 .mf
+		mv .mf Makefile
+	else
+		rm -rf .mf
+	fi
+else
+	chmod 755 .mf
+	mv .mf Makefile
+fi
